requirements.txt
streamlit
google-generativeai
murf
requests
deep-translator
